requirements.txt
streamlit==1.35.0
scikit-learn==1.5.0
pillow>=10.3.0
numpy
pandas
matplotlib
